Fixing favorites subpage
